feat: Basic part of fastjson vulnerability
This article is the first chapter of the `Fastjson` vulnerability basic article of the `Java` security mother-level zero-basic tutorial. It starts from the pre-knowledge, and then transitions to the recurrence of the vulnerability and the analysis of the code. This article removes the code for a total of nearly` 18000 words, 108 pictures, the pictures are detailed and clear enough, following the recurrence analysis, we can basically figure out what these loopholes are all about. The improvement article will focus on the other `payload` of `Fastjson` and the use of `Fastjson` not going online, and will be updated in the next time.